+++ src/main/java/org/example/data_structures/Tuple.java
@@ -1,4 +1,6 @@
-package org.example;
+package org.example.data_structures;
+
+import org.example.exceptions.DBAppException;
 
 import java.util.Hashtable;
