+++ src/main/java/org/example/data_structures/index/BTreeInnerNode.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.data_structures.index;
 
 import java.io.Serializable;
