+++ src/main/java/org/example/data_structures/index/DoubleIndex.java
@@ -1,8 +1,12 @@
-package org.example;
+package org.example.data_structures.index;
+
+import org.example.exceptions.DBAppException;
+import org.example.data_structures.Page;
+import org.example.data_structures.Table;
 
 import java.util.Vector;
 
-import static org.example.FileManager.deserializePage;
+import static org.example.managers.FileManager.deserializePage;
 
 public class DoubleIndex implements Index {
