+++ src/main/java/org/example/data_structures/index/StringIndex.java
@@ -1,18 +1,22 @@
-package org.example;
+package org.example.data_structures.index;
+
+import org.example.exceptions.DBAppException;
+import org.example.data_structures.Page;
+import org.example.data_structures.Table;
 
 import java.util.Vector;
 
-import static org.example.FileManager.deserializePage;
+import static org.example.managers.FileManager.deserializePage;
 
-public class IntegerIndex implements Index {
+public class StringIndex implements Index {
 
-    private String columnType;
-    private String columnName;
-    private Table parentTable;
-    private BTree<Integer,String> bTree;
+    private final String columnType;
+    private final String columnName;
+    private final Table parentTable;
+    private BTree<String,String> bTree;
 
 
-    public IntegerIndex(Table parentTable, String columnType, String columnName) throws IllegalArgumentException {
+    public StringIndex(Table parentTable, String columnType, String columnName) throws IllegalArgumentException {
         this.columnType = columnType;
         this.parentTable = parentTable;
         this.columnName = columnName;
@@ -32,11 +36,11 @@ public class IntegerIndex implements Index {
         return parentTable;
     }
 
-    public BTree<Integer, String> getbTree() {
+    public BTree<String, String> getbTree() {
         return bTree;
     }
-    
-    
+
+
     public void populateIndex() throws DBAppException {
         Vector<String> pageNames = parentTable.getPageNames();
         for (String pageName : pageNames) {
@@ -44,23 +48,23 @@ public class IntegerIndex implements Index {
             if (page.getTuples().isEmpty()) {
                 continue;
             }
-            bTree.insert((Integer) page.getTuples().get(0).getValues().get(columnName), pageName);
+            bTree.insert((String) page.getTuples().get(0).getValues().get(columnName), pageName);
         }
     }
 
     @Override
     public void insert(Object key, String value) {
-        bTree.insert((Integer) key, value);
+        bTree.insert((String) key, value);
     }
 
     @Override
     public String search(Object key) {
-        return bTree.search((Integer) key);
+        return bTree.search((String) key);
     }
 
     @Override
     public void delete(Object key) {
-        bTree.delete((Integer) key);
+        bTree.delete((String) key);
     }
-    
+
 }
