+++ src/main/java/org/example/managers/DeletionManager.java
@@ -1,4 +1,9 @@
-package org.example;
+package org.example.managers;
+
+import org.example.exceptions.DBAppException;
+import org.example.data_structures.Tuple;
+import org.example.data_structures.Page;
+import org.example.data_structures.Table;
 
 public class DeletionManager{
     private DeletionManager() {
